requirements.txt
# ============================================================================
# CORE DATA PROCESSING
# ============================================================================
pandas==2.2.3
numpy==2.2.6

# ============================================================================
# DATA VISUALIZATION
# ============================================================================
matplotlib==3.10.3
seaborn==0.13.2
plotly==5.19.0

# ============================================================================
# MACHINE LEARNING
# ============================================================================
scikit-learn==1.6.1
joblib==1.5.1

# ============================================================================
# WEB APPLICATION
# ============================================================================
streamlit==1.45.1

# ============================================================================
# CHATBOT & AI
# ============================================================================
google-generativeai==0.8.5

# ============================================================================
# ENVIRONMENT & CONFIGURATION
# ============================================================================
python-dotenv==1.1.0

# ============================================================================
# UTILITIES
# ============================================================================
requests==2.32.3
